fixed showcasenav on projects#index and versions#show
