+++ app/controllers/versions_controller.rb
@@ -27,6 +27,9 @@ class VersionsController < ApplicationController
   def show
     @version = Version.find(params[:id])
     @apis = @version.apis
+    @result = @apis.map {|api| {:apiId => api.id, :apiMethod => api.method, :apiEndPoint => api.end_point, 
+      :apiDescription => api.description, :parameters => api.params.as_json, :headers => api.headers.as_json, 
+      :bodies => api.bodies.as_json, :responses => api.responses.as_json}}
   end
 
   def edit
