+++ app/views/projects/index.html.erb
@@ -1,3 +1,1 @@
-<% @projects.each do |project| %>
-  <%= react_component('ProjectIndex', {projectId: project.id, projectName: project.name, versions: project.version_details}) %>
-<% end %>
+<%= react_component('ProjectIndex', {projects: @result}) %>
