+++ app/views/versions/show.html.erb
@@ -1,5 +1,1 @@
-<% @apis.each do |api| %>
-<%= react_component('ApiIndex', {apiId: api.id, apiMethod: api.method, apiEndPoint: api.end_point, 
-  apiDescription: api.description, parameters: api.params.as_json, headers: api.headers.as_json, 
-  bodies: api.bodies.as_json, responses: api.responses.as_json}) %>
-<% end %>
+<%= react_component('ApiIndex', {apis: @result}) %>
